LayoutTests/http/tests/plugins/plugin-javascript-access-allow-all-plugins-expected.txt

Test: Confirm the plugin is present.

Name: WebKit Test PlugIn

Description: Simple Netscape plug-in that handles test content for WebKit

Filename: TestNetscapePlugIn.plugin

Mime Types:

Type: application/x-webkit-test-netscape

Description: test netscape content

Suffixes: testnetscape


Plugin.item() works.

Plugin.namedItem() works.

Type: image/png

Description: PNG image

Suffixes: png


Plugin.item() works.

Plugin.namedItem() works.

PluginArray.item() works.

PluginArray.namedItem() works.

PASS! (Found the netscape test plugin on a non-local file!)
